// src/generated/models/Geek_businessprocessesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Geek_businessprocessesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Geek_businessprocessesstatecode = keyof typeof Geek_businessprocessesstatecode;
export const Geek_businessprocessesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Geek_businessprocessesstatuscode = keyof typeof Geek_businessprocessesstatuscode;

export interface Geek_businessprocessesBase {
  geek_businessprocessid: string;
  geek_description?: string;
  geek_documenturl?: string;
  geek_markdowndetails?: string;
  geek_processid?: string;
  geek_processname?: string;
  importsequencenumber?: string;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Geek_businessprocessesstatecode;
  statuscode?: Geek_businessprocessesstatuscode;
  timezoneruleversionnumber?: string;
  utcconversiontimezonecode?: string;
}

export interface Geek_businessprocesses extends Geek_businessprocessesBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}